requirements.txt
langchain_core==0.3.21
langchain_ollama==0.2.0
langchain_openai==0.2.9
python-dotenv==1.0.1
streamlit==1.40.1
sse_starlette